PR: 데이터 백업시 Celebrate_Cnt, Cheer_Cnt 업데이트 스케쥴러 구현 | feature/user-confirm => BE-develop

See merge request s09-webmobile2-sub2/S09P12A609!126

+++ BackEnd/DreamGream/src/main/java/com/ssafy/dreamgream/domain/post/service/PostService.java
@@ -3,7 +3,6 @@ package com.ssafy.dreamgream.domain.post.service;
 import com.ssafy.dreamgream.domain.member.entity.Member;
 import com.ssafy.dreamgream.domain.member.service.MemberService;
 import com.ssafy.dreamgream.domain.post.dto.request.AchievedPostUpdateRequestDto;
-import com.ssafy.dreamgream.domain.post.dto.request.PostUpdateRequestDto;
 import com.ssafy.dreamgream.domain.post.dto.request.UnAchievedPostUpdateRequestDto;
 import com.ssafy.dreamgream.domain.post.dto.response.PostListResponseDto;
 import com.ssafy.dreamgream.domain.post.dto.response.PostResponseDto;
@@ -22,10 +21,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.transaction.Transactional;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
 
 @Service
@@ -156,23 +153,30 @@ public class PostService {
     }
     //== 리스트 조회, 게시글 조회 끝 ==//
 
-
     public void deletePost(Long postId) {
-        String cheer_key = "cheer_post_" + String.valueOf(postId);
-        Set<String> cheer_members = redisTemplate.opsForSet().members(cheer_key);
-        for(String member : cheer_members){
-            String keyMember = "member_"+member;
-            redisTemplate.opsForSet().remove(cheer_key,member);
-            redisTemplate.opsForSet().remove(keyMember,String.valueOf(postId));
+        Member currentMember = memberService.getCurrentMember();
+        String memberId = String.valueOf(currentMember.getMemberId());
+        String postMemberId = String.valueOf(postRepository.findById(postId).get().getMember());
+        if(memberId != postMemberId){
+            // TODO: 수정 권한이 없을 경우 예외 처리
+        } else{
+            String cheer_key = "cheer_post_" + String.valueOf(postId);
+            Set<String> cheer_members = redisTemplate.opsForSet().members(cheer_key);
+            for(String member : cheer_members){
+                String keyMember = "member_"+member;
+                redisTemplate.opsForSet().remove(cheer_key,member);
+                redisTemplate.opsForSet().remove(keyMember,String.valueOf(postId));
+            }
+            String congrat_key = "congrat_post_" + String.valueOf(postId);
+            Set<String> congrat_members = redisTemplate.opsForSet().members(congrat_key);
+            for(String member : congrat_members){
+                String keyMember = "member_"+member;
+                redisTemplate.opsForSet().remove(congrat_key,member);
+                redisTemplate.opsForSet().remove(keyMember,String.valueOf(postId));
+            }
+            postRepository.deleteById(postId);
         }
-        String congrat_key = "congrat_post_" + String.valueOf(postId);
-        Set<String> congrat_members = redisTemplate.opsForSet().members(congrat_key);
-        for(String member : congrat_members){
-            String keyMember = "member_"+member;
-            redisTemplate.opsForSet().remove(congrat_key,member);
-            redisTemplate.opsForSet().remove(keyMember,String.valueOf(postId));
-        }
-        postRepository.deleteById(postId);
+
     }
 
     @Transactional
